requirements.txt
# Aplicación Web
streamlit>=1.28.0
python-dotenv>=1.0.0

# Manipulación de Datos
pandas>=2.0.0
numpy>=1.24.0

# Visualización
plotly>=5.17.0

# Base de Datos
sqlalchemy>=2.0.0

# Procesamiento de Imágenes
Pillow>=10.0.0

# Requests para APIs
requests>=2.31.0

# Jupyter Notebooks (opcional, para desarrollo)
jupyter>=1.0.0
ipykernel>=6.25.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
